Dash/requirements.txt
dash
dash-renderer
dash-html-components
dash-core-components
plotly
dash-bootstrap-components
websocket-client